Moved PostAsJson to a helper static class
