+++ Http/HttpHelpers.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 567069ad90ee54dc4a9e0dd97e888c76
+timeCreated: 1513171847
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
